+++ build_runner/test/generate/watch_test.dart
@@ -154,8 +154,8 @@ void main() {
 
       test('ignores events from nested packages', () async {
         var writer = new InMemoryRunnerAssetWriter();
-        final packageGraph = buildPackageGraph('a', {
-          package('a', path: path.absolute('a')): ['b'],
+        final packageGraph = buildPackageGraph({
+          rootPackage('a', path: path.absolute('a')): ['b'],
           package('b', path: path.absolute('a', 'b')): []
         });
 
@@ -408,7 +408,7 @@ Stream<BuildResult> startWatch(List<BuildAction> buildActions,
   final actualAssets = writer.assets;
   final reader = new InMemoryRunnerAssetReader(actualAssets);
   packageGraph ??=
-      buildPackageGraph('a', {package('a', path: path.absolute('a')): []});
+      buildPackageGraph({rootPackage('a', path: path.absolute('a')): []});
   final watcherFactory = (String path) => new FakeWatcher(path);
 
   var buildState = watch_impl.watch(buildActions,